requirements.txt
Flask==3.0.3
PyMuPDF==1.24.1
requests==2.31.0
python-dotenv==1.0.1
gunicorn==22.0.0